+++ src/instructlab/model/download.py
@@ -1,17 +1,19 @@
 # SPDX-License-Identifier: Apache-2.0
 
 # Standard
+from abc import ABC, abstractmethod
 from pathlib import Path
-import abc
+import hashlib
+import json
 import logging
 import os
 import re
 import subprocess
 
 # Third Party
-from huggingface_hub import hf_hub_download, list_repo_files
+from huggingface_hub import hf_hub_download
 from huggingface_hub import logging as hf_logging
-from huggingface_hub import snapshot_download
+from huggingface_hub import model_info, snapshot_download
 from packaging import version
 import click
 
@@ -19,12 +21,45 @@ import click
 from instructlab import clickext
 from instructlab.configuration import DEFAULTS
 from instructlab.defaults import DEFAULT_INDENT
-from instructlab.utils import is_huggingface_repo, is_oci_repo, load_json
+from instructlab.utils import load_json
 
 logger = logging.getLogger(__name__)
 
 
-class Downloader(abc.ABC):
+class ModelMetadata:
+    """Definition for metadata that is stored for downloaded models"""
+
+    def __init__(self, v: str = "", SHA: str = "", size: int = 0) -> None:
+        self.version = v
+        self.SHA = SHA
+        self.size = size
+
+
+class DownloadTarget:
+    """Parameters for model being targeted for download"""
+
+    def __init__(
+        self, repository: str = "", filename: str | None = None, release: str = "main"
+    ) -> None:
+        self.repository = repository
+        self.filename = filename
+        self.release = release
+
+
+DEFAULT_TARGETS = [
+    DownloadTarget(
+        DEFAULTS.GRANITE_GGUF_REPO, DEFAULTS.GRANITE_GGUF_MODEL_NAME, "main"
+    ),
+    DownloadTarget(
+        DEFAULTS.MERLINITE_GGUF_REPO, DEFAULTS.MERLINITE_GGUF_MODEL_NAME, "main"
+    ),
+    DownloadTarget(
+        DEFAULTS.MISTRAL_GGUF_REPO, DEFAULTS.MISTRAL_GGUF_MODEL_NAME, "main"
+    ),
+]
+
+
+class Downloader(ABC):
     """Base class for a downloading backend"""
 
     def __init__(
@@ -38,10 +73,25 @@ class Downloader(abc.ABC):
         self.repository = repository
         self.release = release
         self.download_dest = download_dest
+        self.metadata = ModelMetadata()
+        self.metadata_filename = "metadata.json"
 
-    @abc.abstractmethod
+    @abstractmethod
     def download(self) -> None:
         """Downloads model from specified repo/release and stores it into download_dest"""
+
+    @abstractmethod
+    def get_metadata(self) -> None:
+        """Retrieves metadata for model to be downloaded"""
+
+    def dump_metadata(self, dest: Path) -> None:
+        """Dumps metadata for downloaded model onto disk"""
+        try:
+            with open(dest / self.metadata_filename, "w", encoding="utf-8") as f:
+                json.dump(self.metadata.__dict__, f)
+        except OSError as e:
+            # not blocking on failure to dump model metadata to disk
+            logger.error(f"Failed to dump model metadata for {dest}: {e}")
 
 
 class HFDownloader(Downloader):
@@ -61,13 +111,62 @@ class HFDownloader(Downloader):
         )
         self.filename = filename
         self.hf_token = hf_token
+        self.hf_prefix = "huggingface.co"
+
+    @staticmethod
+    def _is_huggingface_repo(repo_name: str) -> tuple[bool, str]:
+        """
+        Checks if a provided repository follows the huggingface URL syntax, and extracts
+        the repository path if it does
+        """
+        # allow alphanumerics, underscores, hyphens and periods in huggingface repo names
+        # repo name should be of the format <owner>/<model> or huggingface.co/<owner>/<model>
+        pattern = r"^(huggingface\.co\/)?([\w.-]+\/[\w.-]+)$"
+        match = re.search(pattern, repo_name)
+        if match:
+            return True, match.group(2)
+        return False, ""
+
+    def get_metadata(self) -> None:
+        self.metadata.version = self.release
+
+        modelInfo = model_info(
+            repo_id=self.repository,
+            revision=self.release,
+            files_metadata=True,
+        )
+
+        # store commit SHA
+        self.metadata.SHA = modelInfo.sha
+
+        # calculate and store size in bytes so it can be used for system operations easily
+        model_size = 0
+        for sibling in modelInfo.siblings:
+            # if a filename is specified take that file's size (gguf model)
+            # else add up sizes of all files in the repo
+            if self.filename is not None and self.filename == sibling.rfilename:
+                model_size = sibling.size
+                break
+            model_size += sibling.size
+
+        self.metadata.size = model_size
 
     def download(self):
         """
         Download specified model from Hugging Face
         """
+        # allow/encourage users to prefix HF model URLs with "huggingface.co" for consistency with other
+        # model source URLs and to align with the storage directory structure, but strip it off before making HF API calls
+        _, self.repository = HFDownloader._is_huggingface_repo(self.repository)
+        self.release = self.release if self.release != "" else "main"
+        self.get_metadata()
+
+        destination = os.path.join(
+            self.download_dest, self.hf_prefix, self.repository, self.release
+        )
+
         click.echo(
-            f"Downloading model from Hugging Face:\n{DEFAULT_INDENT}Model: {self.repository}@{self.release}\n{DEFAULT_INDENT}Destination: {self.download_dest}"
+            f"Downloading model from Hugging Face:\n{DEFAULT_INDENT}Model: {self.repository}@{self.release}\n{DEFAULT_INDENT}Destination: {destination}"
         )
 
         if self.hf_token == "" and "instructlab" not in self.repository:
@@ -80,11 +179,12 @@ class HFDownloader(Downloader):
         try:
             if self.ctx.obj is not None:
                 hf_logging.set_verbosity(self.ctx.obj.config.general.log_level.upper())
-            files = list_repo_files(repo_id=self.repository, token=self.hf_token)
-            if any(re.search(r"\.(safetensors|bin)$", fname) for fname in files):
-                self.download_entire_hf_repo()
+            if self.filename != "":
+                self.download_gguf(destination)
             else:
-                self.download_gguf()
+                self.download_entire_hf_repo(destination)
+
+            super().dump_metadata(Path(destination))
 
         except Exception as exc:
             click.secho(
@@ -93,14 +193,14 @@ class HFDownloader(Downloader):
             )
             raise click.exceptions.Exit(1)
 
-    def download_gguf(self) -> None:
+    def download_gguf(self, destination: str) -> None:
         try:
             hf_hub_download(
                 token=self.hf_token,
                 repo_id=self.repository,
                 revision=self.release,
                 filename=self.filename,
-                local_dir=self.download_dest,
+                local_dir=destination,
             )
 
         except Exception as exc:
@@ -110,16 +210,15 @@ class HFDownloader(Downloader):
             )
             raise click.exceptions.Exit(1)
 
-    def download_entire_hf_repo(self) -> None:
+    def download_entire_hf_repo(self, destination: str) -> None:
         try:
-            local_dir = os.path.join(self.download_dest, self.repository)
-            os.makedirs(name=local_dir, exist_ok=True)
+            os.makedirs(name=destination, exist_ok=True)
 
             snapshot_download(
                 token=self.hf_token,
                 repo_id=self.repository,
                 revision=self.release,
-                local_dir=local_dir,
+                local_dir=destination,
             )
         except Exception as exc:
             click.secho(
@@ -134,6 +233,61 @@ class OCIDownloader(Downloader):
     Class to handle downloading safetensors models from OCI Registries
     We are leveraging OCI v1.1 for this functionality
     """
+
+    def get_metadata(self) -> None:
+        self.metadata.version = self.release
+
+        # Check if skopeo is installed and the version is at least 1.9
+        check_skopeo_version()
+
+        # Run skopeo inspect to retrieve raw OCI manifest
+        command = [
+            "skopeo",
+            "inspect",
+            "--raw",
+            f"{self.repository}:{self.release}",
+        ]
+        logger.debug(f"Running skopeo command: {command}")
+
+        try:
+            result = subprocess.run(command, stdout=subprocess.PIPE, check=True)
+        except subprocess.CalledProcessError as e:
+            click.secho(
+                f"Failed to run skopeo command: {e}.\nstdout: {e.stdout}.\nstderr: {e.stderr}",
+                fg="red",
+            )
+        manifest = result.stdout.decode("utf-8")
+
+        # calculate and store the manifest digest
+        self.metadata.SHA = hashlib.sha256(manifest.encode()).hexdigest()
+
+        # calculate model size from layer information in manifest
+        index = json.loads(manifest)
+        model_size = 0
+        for layer in index["layers"]:
+            model_size += layer["size"]
+
+        # storing size in bytes so it can be used for system operations easily
+        self.metadata.size = model_size
+
+    @staticmethod
+    def _is_oci_repo(repo_url: str) -> tuple[bool, str]:
+        """
+        Checks if a provided repository follows the OCI registry URL syntax, and extracts
+        the repository path if it does
+        """
+
+        oci_url_regex = r"^docker://([a-zA-Z0-9\-_.]+(:[0-9]+)?)(/[a-zA-Z0-9\-_.]+)+(@[a-zA-Z0-9]+|:[a-zA-Z0-9\-_.]+)?$"
+        match = re.search(oci_url_regex, repo_url)
+        if match:
+            # group(0) is the full match, stripping the 'docker://' prefix
+            full_repo_path = match.group(0)[9:]
+            # Strip tag or digest if it exists (after ':' or '@')
+            repo_without_tag = re.sub(
+                r"(:[a-zA-Z0-9\-_.]+|@[a-zA-Z0-9]+)$", "", full_repo_path
+            )
+            return True, repo_without_tag
+        return False, ""
 
     @staticmethod
     def _extract_sha(sha: str):
@@ -188,26 +342,32 @@ class OCIDownloader(Downloader):
         return oci_model_file_map
 
     def download(self):
-        click.echo(
-            f"Downloading model from OCI registry:\n{DEFAULT_INDENT}Model: {self.repository}@{self.release}\n{DEFAULT_INDENT}Destination: {self.download_dest}"
-        )
+        self.get_metadata()
 
-        # raise an exception if user specified tag/SHA embedded in repository instead of specifying --release
-        match = re.search(r"^(?:[^:]*:){2}(.*)$", self.repository)
-        if match:
-            click.secho(
-                f"\nInvalid repository supplied:\n{DEFAULT_INDENT}Please specify tag/version '{match.group(1)}' via --release",
-                fg="red",
+        # Try to isolate the path to where the model is stored via the URL and replicate that structure locally
+        # If unable to do that for whatever reason, just use the model name and create a folder out of that
+        model_sub_dirs = []
+        _, repo = OCIDownloader._is_oci_repo(self.repository)
+        if repo != "":
+            model_sub_dirs = repo.split("/")
+            model_dest_dir = os.path.join(
+                self.download_dest, *model_sub_dirs, self.release
             )
-            raise click.exceptions.Exit(1)
-
-        model_name = self.repository.split("/")[-1]
-        os.makedirs(os.path.join(self.download_dest, model_name), exist_ok=True)
-        oci_dir = f"{DEFAULTS.OCI_DIR}/{model_name}"
+            oci_dir = os.path.join(DEFAULTS.OCI_DIR, *model_sub_dirs, self.release)
+        else:
+            model_name = self.repository.split("/")[-1]
+            model_dest_dir = os.path.join(self.download_dest, model_name, self.release)
+            oci_dir = f"{DEFAULTS.OCI_DIR}/{model_name}/{self.release}"
+        os.makedirs(model_dest_dir, exist_ok=True)
         os.makedirs(oci_dir, exist_ok=True)
 
-        # Check if skopeo is installed and the version is at least 1.9
-        check_skopeo_version()
+        dest_model_path = os.path.join(
+            self.download_dest, *model_sub_dirs, self.release
+        )
+
+        click.echo(
+            f"Downloading model from OCI registry:\n{DEFAULT_INDENT}Model: {self.repository}@{self.release}\n{DEFAULT_INDENT}Destination: {dest_model_path}"
+        )
 
         command = [
             "skopeo",
@@ -238,20 +398,23 @@ class OCIDownloader(Downloader):
             raise click.exceptions.Exit(1)
 
         blob_dir = f"{oci_dir}/blobs/sha256/"
-        for name, dest in file_map.items():
-            dest_model_path = Path(self.download_dest) / model_name / str(dest)
-            # unlink any existing version of the file
-            if dest_model_path.exists():
-                dest_model_path.unlink()
 
-            if not dest_model_path.parent.exists():
-                dest_model_path.parent.mkdir(parents=True)
+        for name, dest in file_map.items():
+            model_file = Path(dest_model_path) / str(dest)
+            # unlink any existing version of the file
+            if model_file.exists():
+                model_file.unlink()
+
+            if not model_file.parent.exists():
+                model_file.parent.mkdir(parents=True)
 
             # create symlink to files in cache to avoid redownloading if model has been downloaded before
             os.symlink(
                 os.path.join(blob_dir, name),
-                dest_model_path,
+                model_file,
             )
+
+        super().dump_metadata(Path(dest_model_path))
 
 
 @click.command()
@@ -260,11 +423,6 @@ class OCIDownloader(Downloader):
     "-rp",
     "repositories",
     multiple=True,
-    default=[
-        DEFAULTS.GRANITE_GGUF_REPO,
-        DEFAULTS.MERLINITE_GGUF_REPO,
-        DEFAULTS.MISTRAL_GGUF_REPO,
-    ],  # TODO: add to config.yaml
     show_default=True,
     help="Hugging Face or OCI repository of the model to download.",
 )
@@ -273,11 +431,6 @@ class OCIDownloader(Downloader):
     "-rl",
     "releases",
     multiple=True,
-    default=[
-        "main",
-        "main",
-        "main",
-    ],  # TODO: add to config.yaml
     show_default=True,
     help="The revision of the model to download - e.g. a branch, tag, or commit hash for Hugging Face repositories and tag or commit hash for OCI repositories.",
 )
@@ -285,11 +438,6 @@ class OCIDownloader(Downloader):
     "--filename",
     "filenames",
     multiple=True,
-    default=[
-        DEFAULTS.GRANITE_GGUF_MODEL_NAME,
-        DEFAULTS.MERLINITE_GGUF_MODEL_NAME,
-        DEFAULTS.MISTRAL_GGUF_MODEL_NAME,
-    ],
     show_default="The default model location in the instructlab data directory.",
     help="Name of the model file to download from the Hugging Face repository.",
 )
@@ -310,27 +458,53 @@ class OCIDownloader(Downloader):
 def download(ctx, repositories, releases, filenames, model_dir, hf_token):
     """Downloads model from a specified repository"""
     downloader = None
+    targets = []
 
-    # strict = false ensures that if you just give --repository <some_safetensor> we won't error because len(filenames) is greater due to the defaults
-    for repository, filename, release in zip(
-        repositories, filenames, releases, strict=False
-    ):
-        if is_oci_repo(repository):
-            downloader = OCIDownloader(
-                ctx=ctx, repository=repository, release=release, download_dest=model_dir
+    # fill in defaults only when user passes no arguments to `ilab model download`
+    if not repositories and not filenames and not releases:
+        targets = DEFAULT_TARGETS
+    else:
+        for idx, repo in enumerate(repositories):
+            targets.append(
+                DownloadTarget(
+                    repo,
+                    filenames[idx] if idx < len(filenames) else "",
+                    releases[idx] if idx < len(releases) else "",
+                )
             )
-        elif is_huggingface_repo(repository):
+
+    for tgt in targets:
+        # raise an exception if user specified tag/SHA embedded in repository that conflicts with --release
+        match = re.search(r"(.*):(\w+)$", tgt.repository)
+        if match:
+            if tgt.release != "" and tgt.release != match.group(2):
+                click.secho(
+                    f"Conflicting versions supplied: '{match.group(2)}' and {tgt.release}. Pkease specify one or the other.",
+                    fg="red",
+                )
+                raise click.exceptions.Exit(1)
+            tgt.release = match.group(2)
+            tgt.repository = match.group(1)
+
+        if OCIDownloader._is_oci_repo(tgt.repository)[0]:
+            downloader = OCIDownloader(
+                ctx=ctx,
+                repository=tgt.repository,
+                release=tgt.release,
+                download_dest=model_dir,
+            )
+        elif HFDownloader._is_huggingface_repo(tgt.repository)[0]:
             downloader = HFDownloader(
                 ctx=ctx,
-                repository=repository,
-                release=release,
+                repository=tgt.repository,
+                release=tgt.release,
                 download_dest=model_dir,
-                filename=filename,
+                filename=tgt.filename,
                 hf_token=hf_token,
             )
         else:
             click.secho(
-                f"repository {repository} matches neither Hugging Face nor OCI registry format.\nPlease supply a valid repository",
+                f"repository {tgt.repository} matches neither Hugging Face nor OCI registry format.\nPlease supply a valid repository",
                 fg="red",
             )
             raise click.exceptions.Exit(1)